Updated openstack/openstack Project: openstack/trove  2faed4f8fb4e6953fc33acbdc3aa586b8c788230
Move mysql datadir to a sub-directory on mounted volume

Having the mysql datadir located at root of a mounted volume
(which happens with volume support enabled) causes problems for
backup/restore and resize. Mysql decides that the "lost+found"
directory, which the operating system puts at the root of each
file system, is actually a mysql "database".

This change causes the guest agent to move the datadir into a
sub-directory called "data" on the mounted volume.

Change-Id: Ica628012a5708374d73e5394e370da2514300939
Co-Authored-By: Simon Chang <schang@tesora.com>
Closes-bug: 1370646