Updated openstack/openstack Project: openstack-infra/devstack-gate  5a9565c8f746f4d514a2127860c79f5485a2c1bd
Unmount the ephemeral drive before formatting it

HPCloud is instructing cloud-init to mount the ephemeral drives now via
metadata. This means that the disk will be mounted by the time d-g runs
previously resulting in a failed job due to the return 1. Instead of
erroring write a warning then unmount the ephemeral drive and format it.

Change-Id: I1c27e6728cccd13aaa03d782a7f6c27065e3447b