Updated openstack/openstack Project: openstack-infra/system-config  f68f08620a7dfbd735888746be9085b778f39049
Root the Change-Id and commit commentlinks

Gerrit's query redirects have changed slightly since 2.8, and it's
possible to end up browsing with a URL like https://hostname// which
causes relative-path links such as the Change-ID or commit SHA
commentlinks to 404. Prepend a leading / to the commentlink for this
and avoid the issue.

Change-Id: I09f0b491e2f54ab10d29f456cd026d90b673447f