Updated openstack/openstack Project: openstack-infra/system-config  d165b69d372beb8e291571b87b448d33c2c59601
Use profile in clouds.yaml instead of cloud

The cloud moniker was confusing to people, so we added support in
os-client-config for calling it profile instead. Update the clouds.yaml
to indicate this. Also, change hp and tripleo to project_name since
that's what those are. Rackspace is still a project_id.

Change-Id: I77b2153a096ba3fced2e4d7c1f63663f798d4046