Updated openstack/openstack Project: openstack-infra/system-config  5956ad58d929830393c20bb31ff232c4be40e361
Symlink yaml2ical calendar folder to apache dir

The DocumentRoot for eavesdrop needs a symlink to the calendars dir
now that yaml2ical places individual icals in there.

For this to work we need to wait for the calendar dir to be created
from a successful run on the publisher job.

Change-Id: Icd6fd2969fcea3ef9b7ecd7d482941dbd92b2d26
Depends-On: Ic968a1883d3942335547ed52948030f7f1f84add