Updated openstack/openstack Project: openstack/oslo.versionedobjects  927014c451076c8db8a58600b8f372a73c7d86e8
Fix serializer supported version reporting in object_backport

The serializer needs to report the version of the toplevel object
that we were trying to deserialize when we encountered the incompatible
version. We depend on the implementor of the indirection_api to use
the object relationship mappings to return to us a fully-supported
object tree based on the version at the top.

This patch fixes the serializer to report the toplevel version instead of
the one that caused the fault.

Change-Id: I7ac03717e17cd7229abfeef6d9c7d9d8ec9be27a
Closes-Bug: #1475356