Updated openstack/openstack Project: openstack/nova  9b8891de4b9ce8a1cec153bec9a4723d695e4305
Fix the incorrect PciDeviceList version number

The following commit missed to bump the related list when bumping the
object version.

    commit 1d4ad44f15503b8710ff59dea9f5e3f4894d56fb
    Author: Przemyslaw Czesnowicz <przemyslaw.czesnowicz@intel.com>
    Date:   Thu Nov 20 12:04:05 2014 +0000

        Add numa_node to PCIDevice

        The change extends the PCI device model and its DB representation.
        Query libvirt for the PCI devices numa node.

This patch adds the necessary changes to master. We only bump the
PciDeviceList version, and add a backporting rule to obj_relationship
dictionary of the Instance object, to start sending the new version with
1.20 version of the instance (we are currently on 1.21).

Landing this patch fixes versioning going forward, and all interactions
that request the PciDeviceList through the Instance object, however
applying the stable/kilo backport of this is required for Kilo nodes to
start requesting the latest version of PciDevice through PciDeviceList.

We can't send 1.2 PciDeviceList to Kilo nodes yet because of
related-bug 1475254 , so we have to backport this to Kilo first before
we can start sending the 1.2 to Kilo nodes.

Related-bug: #1474074
Change-Id: Id743848061d2bcd64d785979c7a11ee7c945f06b