Updated openstack/openstack Project: openstack/cinder  2138ff314adea98a6f5eb3b8d6a64f3c402ef336
Handle volume not found on zfssa volume delete

The zfssa driver doesn't handle cases where a volume may not
exist on the backend when a delete call is sent.  The result
is a stack-trace in the logs, sometimes a crash of the service
and an inability to clean up volumes from Cinder's perspective.

This patch doesn't address the root cause of the issue, but
it does handle the case more gracefully by catching the
exception, logging an error and returning.

This way the volume can be cleaned up on the Cinder side and
the operator still has an indication that something went wrong.
This is a common pattern for most of the drivers in Cinder.

Change-Id: I09725b29effb79450d010949527bd54329919f52
Closes-Bug: #1472412