Updated openstack/openstack Project: openstack/cinder  a5c988f93eda5581899ad67c514ed1cb5666ac27
Fix getting out-of-date volume operation state issue for VNX

The driver tries to get a volume's operation state with no polling
in the function wait_until_lun_ready_for_ops, that means the driver
may get out-of-date data. For example, if a volume has just changed
to 'Preparing' state, the function may get the state as 'Ready' and
report that the volume is ready to use. The fix is to get the volume
state with polling at the first time, then the driver would report
the right state of the volume.

Change-Id: I07aa89c33df4935716dbb35d3cfcdbc0e9964e0e
Closes-Bug: #1470041