Updated openstack/openstack Project: openstack/tempest  94131ba8bf5c544fd84607870a47ab2a69631439
Ensure we close the file accounts file after reading

This commit fixes a small bug in the accounts.yaml file reading.
Previously we were opening the file but never closing it. This was
probably safe because we're tempest is read only on the file nothing
else is normally trying to use it, that and python would take care of
the close when things exit. But, it's better to make sure we do things
correctly and close the file after we're done with it. This commit does
just that.

Change-Id: I8a7a360c8908160b07e448bce8af33c7f528815e