Updated openstack/openstack Project: openstack/trove  834e1d64d871aece0f21c203582875070480ec53
Support authentication in the MongoDB guest agent

NOTE: The code to handle MongoDB authentication is being added, but the auth
switch will NOT be turned on in this commit.

Implement client authentication in the MongoDB guest agent. This involves
creating an os_admin user, authenticating the client before all requests,
and passing a keyfile to new cluster instances.

A MongoDB client driver is being added to improve request handling.

The os_admin password and the keyfile key will be randomly generated and
stored on the guests.

This work should be transparent to the user. The result of Trove instance
and cluster creation should not be effected.

New classes:
* MongoDBSchema - represents Mongo databases in Trove
* MongoDBUser   - represents Mongo users in Trove
* MongoDBAdmin  - wrapper for running administrative commands
* MongoDBClient - wrapper for the PyMongo client driver
* MongoDBCredentials - utility class for handling username/password

Test:
* Added 'pymongo' to test-requirements.txt
* Refactored the test_mongodb_manager unit test to make it easier to add
  new tests.
* Cleaned up dangling mocks.
* Added test for the 'secure' procedure during 'prepare'.

RELATED CHANGES:

Upcoming patch sets that enable Trove management of MongoDB databases and
users require these changes. See the dependancy tree below.

--\ authentication
   --\ users create/list/show/delete
      --- users root enable/show
      --- users access grant/show/revoke
   --- database

Closes-Bug: #1467513
Partially Implements: blueprint mongodb-database
Change-Id: I90793f95984b360a2c3668421cb5351aac89e46f