Updated openstack/openstack Project: openstack-infra/system-config  f889e14fdb8f5bf8324286fec5d92ee39b7e1916
Add a common data directory for job workers

It's useful to be able to add data files on our workers as a common
reference, much in the same way job scripts are provided. This will
be used to keep a local copy of the bindep fallback package list so
that it doesn't require directly cloning the project-config repo
during job run-time.

Change-Id: I4487b841f596cdfa2535c944aa3efe32a501b3c9
Depends-On: I30a8be9167cc5c52faae59a35a284b3f37b93f44