Updated openstack/openstack Project: openstack/cinder-specs  053f4ebd53f1ad2bc3d2182463b8df8328dbbc45
Incremental backup improvements for L

In Kilo release we supported incremental backup, but there are still some
points need to be improved.
1. From the API perspective, there is no place to show the backup is
incremental or not. It's useful to 3rd part billing system, they could
have different fee for full and incremental backup.
2. User also doesn't know if the incremental backup can be deleted or
not, it's important because cinder doesn't allow deleting this backup since
'Incremental backups exist for this backup'. Right now, they must have a try
to know it. So if there is a flag to indicate the backup can't be deleted
or not, it will bring more convenience to user and reduce API call.
3.Enriching the notification system via reporting to Ceilometer, add
parent_id to report.

blueprint cinder-incremental-backup-improvements-for-l

Change-Id: I39e12c213a39ef28f3523f25f48cae4348678c39