Updated openstack/openstack Project: openstack-infra/puppet-log_processor  17883b76b07108ce289e0fb8b10081e7beac5f7d
Import socket so we can use it to get name info

Previously this was using socket.getaddrinfo() without importing socket
and causing the daemon to fail. Running in the foreground did not use
statsd thus did not attempt to resolve the statsd host which is how this
got past manual testing. Import socket to get everything working agian.

Change-Id: I280973bdcdf472736a07d19173559b062ed74d3c