Updated openstack/openstack Project: openstack/taskflow  a289c7679c86b12580815da2d5fc689d4a857839
Run the '99_bottles.py' demo at a fast rate when activated

When this is ran without any arguments just run it locally
by starting a producer, then a conductor, then stopping after
it finishes the first song request.

This allows this example to be ran during unit testing to make
sure it functions as expected (with zero return code).

Change-Id: I26c210e2c993e770955985c9c779d303eb8c0616