Updated openstack/openstack Project: openstack-infra/zuul  879dafb5b910b4c8e9fc37895495b4ddf2d58ef5
Reset repos with files named 'HEAD'

If a git repo has the misfortune to have a file with the name "HEAD"
at the root level of the repo, git will return an error because it
is unsure whether the file or ref is meant.

Call 'git reset' with '--' at the end of the command to disambiguate
and instruct git that the HEAD ref is meant.

Change-Id: Icb19063afafa1e0719b3354470c14df8168a64b5