Updated openstack/openstack Project: openstack-infra/system-config  785fbb2b584691bbba4235c009f36abff43884ed
Increase git haproxy maxconn from 32 to 256

It seems from the logs that we're regularly hitting the max 32
connections limit in haproxy for the git protocol backend under
periods of moderate to high load. Increase this to 256, and also
increase the backlog to 256 so that we can have up to one backlogged
connection per active connection.

In theory 256 should be a sane max since we have leastconn balancing
across 8 backends running persistent git-daemon with its default max
connections value of 32.

Change-Id: I80785348b8abb831a2f833a535a5e4234ce3e7f3