Updated openstack/openstack Project: openstack-infra/zuul  6bc782de75be6720aa31843e64f4cb57bfe01de9
Handle adding a job to a failing change during reconfig

During live reconfiguration, if an item was either in a merge
conflict state, or if it had jobs marked as skipped because of
a failing superior job in a job tree, any new jobs added to that
item during the reconfiguration would not have their state
updated to 'skipped'.  This would cause the item to stay in the
queue indefinitely.

Correct this by performing again the state updates that set build
statuses to 'skipped' after an item is re-enqueued.

Change-Id: I96195cb9996b01075e3705b6cd89a9863528898e