Updated openstack/openstack Project: openstack/diskimage-builder  148a5fc40a2d95430aa0f59eb6026923d310c992
Add YUM variable to for Fedora >= 22

Add a YUM variable that defaults to dnf for Fedora 22 and greater.  At
this stage the yum element can do double-duty with dnf -- it's mostly
the same.  If we find it starts getting too unwieldy we can separate
this later.

Modify the install-packages for yum to use this variable when set, but
default back to "yum" to retain the status-quo.

Change-Id: Ibff71465b392d9f66b6f93955ff9223575d6165c