Updated openstack/openstack Project: openstack/neutron  d1f768e064fe3c0957a2b51d3ab2b673c2933b40
Add sanity_check for keepalived ipv6 support

Currently Neutron does not validate the version of Keepalived.
In order to support configuring IPv6 default route, the minimum
version [1] of keepalived that is required is 1.2.10
This patch validates the required support in keepalived.

Although keepalived changelog mentions that IPv6 virtual_routes and
static_routes are supported in version 1.2.8, it is seen that in
v1.2.8 IPv6 default route is not supported. Support for default route
is added in keepalived v1.2.10. Observations on v1.2.8 are captured
at the following link - http://paste.openstack.org/show/165403/

[1] - http://www.keepalived.org/changelog.html

Closes-Bug: #1415756
Change-Id: I768ae233d2c2e24df93a4736ef6d8f4005770fa5