Updated openstack/openstack Project: openstack/oslo.messaging  431f22cfa5284c00cc445d4c038f7232d2c2393b
Fix amqp connection pool leak in ConnectionContext

When failed in connection.reset(), the current code simply
discards the broken connection, without returning a new one
to the pool, nor adjust pool counter. It results in a connection
leak and eventually blocks the thread.

It is fixed by returning a new connection into the pool.

Change-Id: I2b2c23def718d8f2409f9fc415441ac88d40f5b9
Closes-Bug: #1474698