Updated openstack/openstack Project: openstack/neutron  79e8a219eb203823fabf251e1e62f415f652a8be
Add oslo db retry decorator to non-CRUD actions

The previously added decorators to the create and update handlers
in the API layer only applied to actions that followed the standard
create/update path. However, for API operations like add_router_interface,
a different path is followed that wasn't covered by a retry decorator.

This patch adds the decorator to handle deadlocks in those operations as
well.

Closes-Bug: #1475218
Change-Id: Ib354074e6a3f68cedb95fd774f905d94ca16a830