Updated openstack/openstack Project: openstack/requirements  b2788ceb19d93c558a033d48b47c628cb2817b25
Revert "Move integration test dependencies into bindep"

This reverts commit ce6ef0fe230ff939ae4d583a2262a16d598c32b9

We're only installing bindep on diskimage-builder built images, and
the devstack-trusty images in rax regions are snapshot-based. We'll
need to temporarily revert this until devstack-trusty is switched to
DIB everywhere, or DevStack-based jobs move to the ubuntu-trusty DIB
images we have available in rax currently.

Change-Id: I69c923297c3202fd35b8892ddfa4d8dbf272bf16
Closes-Bug: #1475880