Updated openstack/openstack Project: openstack/training-guides  56669b506dda09b1363bc1b7240e8e76269ff83e
Wait for floating IP to turn up

Since Juno, the floating IP often takes a long time to become pingable.
Hopefully, this will be fixed, but for the time being we just ping the
floating IP until we get a reply (or we reach a time limit and give up).

This patch makes the test script wait longer for the floating IP.

Change-Id: I8838aa7dc9b29274ea98401ce8d3ba3033a0dd0d