Updated openstack/openstack Project: openstack-infra/project-config  8f438090a6890c8b500f4363b7ff6075d4f2a2da
Rearrange pipelines for neutron-vpnaas jobs

Currently the neutron-vpnaas repo uses the integrated-gate-neutron
template. This template includes the gate-grenade-dsvm-neutron job
in the check queue. Right now the grenade job for this repo does
not build. The original intention for adding the grenade job to
this repo was to have the job in the experimental queue.

This patch removes the use of the integrated-gate-neutron
template from the neutron-vpnaas repo. It adds all the jobs used
in the template back to the repo in the pipelines used in the
template with the exception of the gate-grenade-dsvm-neutron job
which is now in the experimental pipeline.

Change-Id: I580864201107357b94296bf20bc3bd6a814f3cd8