Updated openstack/openstack Project: openstack/trove  f7aace809fab77f1ae2304d754495fa54d2f8ce8
Switch to oslo.service

oslo.service has graduated, so trove should consume it.
Remove a "ticks_between_runs" parameter from the periodic tasks.
All periodic_tasks will be executed with the same interval with the value of
the "report_interval" option.
According to the old default value of the "report_interval" option and
the value of the ticks_between_runs paramentr (3), increase
the "report_interval" option to the 30 seconds.
Rename the "exists_notification_ticks" option to the "exists_notification_interval"
and increase the default value of it to 3600, according to the old value of
the "report_interval" option.

DocImpact
The option name and default value of exists_notification_ticks was changed to
exists_notification_interval with 3600 sec. And the new default value of
the report_interval option is 30 sec. The deployers need to make sure that
the new values in the seconds, not in the ticks.

Change-Id: Id14d28146f677faf017160ac93289fd119674fc4
Depends-On: Ia5b887e69853f39b387d309831fb7ce51b881149
Closes-Bug: #1466851