Updated openstack/openstack Project: openstack/trove-integration  c2508c60ce75ceca04f3a9d636a41be2e4266976
Added better error msg if working on Trove branch

If a Trove branch is checked out without having TROVE_BRANCH set,
redstack fails with an error message.  This leads one to believe
that you cannot work on a branch in Trove when using redstack, when
all that needs to be done is to set TROVE_BRANCH=<current_branch>

The error message now prints out a useful hint.

The messages were also modified so that it prints in blue when
switching to the branch only if the branch is different.

Change-Id: Ib21403ce0d136d65aaf5bb435e338195164ea459