Updated openstack/openstack Project: openstack/nova  4e51b1a429b81fade7a2bbaf90e73340ed2dbad9
rebuild: fix rebuild of server with volume attached

This was meant to be fixed by I017bf749f426717dc76cf99a387102848fb1c541 ,
but it didn't take into account that BDM entry was destroyed, which
caused the rebuild to fail when spawning the instance.

Add a new parameter to detach_volume() to bypass destroying of BDM,
as we just want to detach a volume first and then re-attach it again.

A Tempest test is added in I50557c69b54003d3409c8e977966f5332f4fe690
to make sure this is actually tested in the gate.

Closes-Bug: #1440762

Co-Authored-By: melanie witt <melwitt@yahoo-inc.com>

Change-Id: I9134fbf5ce72c32cca91de90001c09e00b4e19e8