Updated openstack/openstack Project: openstack/nova  c4defdb168645298ba52033a659736d623622c53
Fixes "Hyper-V destroy vm fails on Windows Server 2008R2"

Delete vm fails on Windows Server 2008R2 with the following error:
[Error 145] The directory is not empty: 'C:\\OpenStack\\Instances\\instance-00000003'

This is happening because on Windows 2008R2 it takes a while (1-2 sec. )
for Hyper-V to delete the instance specific files that it stores in the
instance folder.

This patch fixes this bug by adding a retry in the delete method if
it gets that specific error: WindowsError 145

Closes Bug: 1467409
Change-Id: I8c194320dfa0c9a00ede3cd5bf38446f896156e3