Updated openstack/openstack Project: openstack/sahara-specs  54b2423f0dad9556e77f4f0b014ac0097f00db2b
Store trusts for all clusters in DB

As per team meeting on 7/9/2015, in order to successfully cleanup
stale clusters from a periodic task, we must store trust ids in the
database until cluster activation in the case of long-running
clusters. This spec update formalizes that plan and adds
commentary on the alternative of in-memory (context) trust_id
storage.

Change-Id: If273d99a320f08b22cc0663ac369cc06c13206d8
Addresses-bug: #1468722