Updated openstack/openstack Project: openstack-infra/zuul  b4c3e66d9e4a181b11e69eb5e06413b835c0f7c1
Gear: bind to a specified address

Along with gear's I884e0ddcaac9c7de35b913dc967a47ee854d0a8a, this is
needed to make sure that the local Gear server listens on a correct
interface -- it should be the same one as we use for contacting it later
on.

FIXME: add a dependency on new enough Gear once it's merged & released.

Change-Id: Ia5a0d098f57dcdb3a730c43ab56e55f61127afe8