Updated openstack/openstack Project: openstack-infra/zuul  c051e5970b0419970b388d3b7542e20f82c32287
Update default merge failed error to mention CRD

The merge failure message is the same whether your change failed to
merge or it was one of its dependencies, including cross-repo
dependencies. Update the default error message for this to mention
the latter possibility so that contributors don't unnecessarily
rebase a dependent change which doesn't actually need it.

Change-Id: Ia17d9ab2a842574cf684b4ebb955dd2ec6e9ec3d