Updated openstack/openstack Project: openstack-infra/project-config  f8352d1b100173292042244ba462000b5bff6304
Add ceph devstack job to glance_store

A regression was introduced to glance_store which affected its
use with ceph.  Ceph jobs are run for several projects, but not
glance_store, therefore the regression was not noted.  Run the
ceph job on glance_store changes, with glance_store installed
from git.

Related-bug: #1475811
Change-Id: Ia95fc01a1e4056babb191429a9289f5c71b5ecfa