Updated openstack/openstack Project: openstack/taskflow  09f95cf10ab845589233bd6554d3a792837f8cc1
Use encodeutils for exception -> string function

The oslo.utils library now provides a better version of
this that always returns a unicode exception message, so
update our usage to use it (and remove our own local
function). This guarantee of unicode also means we have to
update a few other places to make sure we get back bytes
or unicode as needed.

Change-Id: I924380408aaf6d2aec418ceaaf623c75900268f7