Updated openstack/openstack Project: openstack/tooz  f674ed55ce52e933e9a15a9b966569f12517dd19
Use a lua script(s) instead of transactions

Lua scripts are guaranteed to be atomic so we can avoid
transactions (and the resulting potential retries that
can be triggered); in general they are faster than
client side retries and are seemingly preferred over
transactions (transactions also don't rollback so they
aren't really offering to much anyway).

This replaces some of the calls that would benefit from
the atomic nature of scripts with scripts (vs there prior
usage of transactions).

Useful links:

- http://redis.io/commands/eval
- http://redis.io/topics/transactions#redis-scripting-and-transactions
- http://redis.io/topics/transactions#why-redis-does-not-support-roll-backs

Change-Id: Icba07d0c73a283c1b3575c388883bb9a9621322c