Updated openstack/openstack Project: openstack/nova  00f85052b53429d5ce6080fc4214a97584374a57
Fix libguestfs failure in test_can_resize_need_fs_type_specified

The above test was passing a tempfile.NamedTemporaryFile object where
a path was expected. This currently works only coincidentally, because
the vfs falls back to localfs when libguestfs isn't available, and the
value is never actually used. However, it fails when libguestfs is
available in the unit test environment, because libguestfs checks the
types of its arguments.

Change-Id: I1f303079190b6a170da89ad7306926ac5000adde