Updated openstack/openstack Project: openstack-infra/project-config  480db742a28bbe71b43c3e72eefd12410c850d8d
Make neutron-dvr job voting at the gate

The single node DVR job seems to be stable now
and following the neutron full job. So it is
time to move the dvr job to voting job.

http://goo.gl/4YK3oZ

This would allow the other patch owners to consider
the dvr in their design considerations and would
substanstially reduce the bugs introduced in
regression.

Change-Id: I16457cf467e5883ae0932076b170de4c067fe7b6