Updated openstack/openstack Project: openstack-infra/project-config  8ac96d49ce34bb403bdf8c2b51b7b4a4881b4852
Remove gate-tempest-dsvm-docker from nova's check queue

The gate-tempest-dsvm-docker job fails most of the time on nova changes
and nobody from the nova-docker project is monitoring the job, so the
nova team wants it removed from the check queue for nova changes.

This was discussed at the nova liberty midcycle meetup.

Change-Id: Ic8eae3b73601518616f9a32236663865024e76cd