Updated openstack/openstack Project: openstack/nova  a96fdf789a812b451cc928d39755edff1b530058
Decouple instance object tests from the api fakes module

The instance tests were written back when the only non-db model of an instance
we had was in the api fakes module. That module is problematic in several ways,
but also needs to change to be non-db-focused soon. This patch converts
the instance object tests to use fake_db_instance to break it away from the api
fakes module.

Related to bp/drop-flavor-compat

Change-Id: Ib790dce76394ce64e9a0faed0f51cce08d256cd5