Updated openstack/openstack Project: openstack/ironic  57c3065eb8123ca9c9dc50c98c6d8fbb45920e3d
Don't do a premature reservation check in the provision API

This leads to error 409 that have to be retried in client.
Instead, only check for reservation if we can't fulfill the request
due to invalid state. That will keep backward compatibility
while using task_manager retries for node locked case.

Change-Id: Iab0a866492dd41d14ed0385c933e8410073694db