Updated openstack/openstack Project: openstack-infra/release-tools  ce5bb38d67d0fc817b3ffb1c465678511598cd52
Prototype script to release from the new YAML files

The schema of the YAML files exposes an issue with the assumptions in
our current release script, which assumes one repository and one tag.
We'll need to fix that, but in the mean time this is the script I used
to generate the oslotest 1.5.2 release today.

Change-Id: Ic253bb34e367fc3c3c891f06e20575209fc49ad1