Updated openstack/openstack Project: openstack/python-ironicclient  66f83811a7452ac6a90916a185aa3143b71448bc
Allow specifying a set of fields of the Port and Chasis resources

This patch add a "--fields" parameter to the "port-list", "port-show",
"chassis-list", "chassis-show" and "chassis-node-list" commands which the
user can specify a subset of fields that will be returned by the server.

This patch also enhances the error message for check_for_invalid_fileds()
method to show what are the valid fields.

Change-Id: Iaf90d615c9fa68f24e54c23d708ff9c5bee7c46c